fix(contracts): resolve all 8 clippy warnings across 3 contracts
- Removed 5 unnecessary u32-as-u32 casts (list.len() already returns u32)
- Replaced assert!(len >= 1) with assert!(!is_empty()) for clarity
- Added #[allow(clippy::too_many_arguments)] on place_order (8 params
  is by design for the contract API)

cargo clippy --all-targets -- -D warnings now passes cleanly.

+++ src/contracts/limit-order/src/lib.rs
@@ -139,6 +139,7 @@ impl LimitOrder {
     }
 
     /// Place a new limit order. Owner must authorize.
+    #[allow(clippy::too_many_arguments)]
     pub fn place_order(
         env: Env,
         owner: Address,
@@ -325,7 +326,7 @@ impl LimitOrder {
             .get(&DataKey::OrderList)
             .unwrap_or_else(|| Vec::new(&env));
 
-        let total = list.len() as u32;
+        let total = list.len();
         let mut idx = cursor;
         if idx >= total {
             return (Vec::new(&env), None);
